Restructure directory and datasets
All ecl files on the main directory with apt names. CSV datasets in
data folder. These files must be sprayed in prior to using any of the
ecl scripts. A Datasets.ecl file is maintained to easily input data
and not write records of the data over and over again.

See Datasets.ecl and data folder for clarification